Resync with upstream onnxruntime-rs (#11)
* Merge upstream

* Fix errors stemming from merge

* Upgrade tracing-subscriber to resolve security audit issues

* Upgrade zip to resolve security audit issues

* Update bindings for macos

* Add aarch64 to build script

* Fix integration test

* Fixed GPU builds

* Fix verification of GPU build

* Add x86_64 bindings

* Update README to note that arm64 macos is fully supported

+++ onnxruntime/src/tensor/ort_tensor.rs
@@ -53,7 +53,7 @@ where
 
         let shape: Vec<i64> = array.shape().iter().map(|d: &usize| *d as i64).collect();
         let shape_ptr: *const i64 = shape.as_ptr();
-        let shape_len = array.shape().len() as u64;
+        let shape_len = array.shape().len();
 
         match T::tensor_element_data_type() {
             TensorElementDataType::Float
@@ -77,7 +77,7 @@ where
                         ort.CreateTensorWithDataAsOrtValue.unwrap()(
                             memory_info.ptr,
                             tensor_values_ptr,
-                            (array.len() * std::mem::size_of::<T>()) as u64,
+                            array.len() * std::mem::size_of::<T>(),
                             shape_ptr,
                             shape_len,
                             T::tensor_element_data_type().into(),
@@ -129,7 +129,7 @@ where
                         ort.FillStringTensor.unwrap()(
                             tensor_ptr,
                             string_pointers.as_ptr(),
-                            string_pointers.len() as u64,
+                            string_pointers.len(),
                         )
                     })
                 }
@@ -199,7 +199,7 @@ mod tests {
     use crate::{AllocatorType, MemType};
     use ndarray::{arr0, arr1, arr2, arr3};
     use std::ptr;
-    use test_env_log::test;
+    use test_log::test;
 
     #[test]
     fn orttensor_from_array_0d_i32() {
